Fixed handling of multiple Pandoc defaults files
