+++ .github/workflows/main.yml
@@ -40,10 +40,10 @@ jobs:
         run: |
           poetry run pip install --upgrade pip
           poetry install
-      - name: Install supported pandoc 2.17.0.1
+      - name: Install supported pandoc 2.18
         run: |
-          wget https://github.com/jgm/pandoc/releases/download/2.17.0.1/pandoc-2.17.0.1-1-amd64.deb
-          sudo dpkg -i pandoc-2.17.0.1-1-amd64.deb
+          wget https://github.com/jgm/pandoc/releases/download/2.18/pandoc-2.18-1-amd64.deb
+          sudo dpkg -i pandoc-2.18-1-amd64.deb
       - name: Install unsupported pandoc 2.10
         run: |
           wget https://github.com/jgm/pandoc/releases/download/2.10/pandoc-2.10-linux-amd64.tar.gz
