+++ pelican/plugins/pandoc_reader/test/defaults_files/standalone_true.yaml
@@ -1,10 +1,10 @@
-# selfcontained_true.yaml
+# standalone_true.yaml
 #
-# A test default file that sets self-contained to true
+# A defaults file that sets standalone to true which is invalid.
 #
 reader: markdown+smart
 writer: html5
-self-contained: true
+standalone: true
 
 html-math-method:
   method: mathjax
