+++ pelican/plugins/pandoc_reader/test/defaults_files/to_writer_both_given.yaml
@@ -1,8 +1,9 @@
-# valid_defaults.yml
+# to_writer_both_given.yaml
 #
-# A test default file that is valid
+# A defaults file that sets both the to and writer attributes which is invalid.
 #
 reader: markdown+smart
+to: html5
 writer: html5
 
 html-math-method:
