+++ pelican/plugins/pandoc_reader/test/markdown/valid_content_with_citation.md
