+++ pelican/plugins/pandoc_reader/test/test_invalid_metadata.py
@@ -1,24 +1,26 @@
-"""Tests using invalid metadata for pandoc-reader plugin."""
+"""Test using invalid metadata with the pandoc-reader plugin."""
 import os
 import unittest
 
+from pelican.plugins.pandoc_reader import PandocReader
 from pelican.tests.support import get_settings
 
-from pandoc_reader import PandocReader
-
 DIR_PATH = os.path.dirname(__file__)
-TEST_CONTENT_PATH = os.path.abspath(os.path.join(DIR_PATH, "test_content"))
+TEST_CONTENT_PATH = os.path.abspath(os.path.join(DIR_PATH, "markdown"))
 
-# Test settings that will be set in pelicanconf.py by plugin users
-PANDOC_ARGS = ["--mathjax"]
+# These settings will be set in pelicanconf.py by plugin users.
+# Appending --wrap=None so that rendered HTML5 does not have new lines (\n)
+# which causes tests to fail.
+# See https://pandoc.org/MANUAL.html#general-writer-options
+PANDOC_ARGS = ["--mathjax", "--wrap=none"]
 PANDOC_EXTENSIONS = ["+smart"]
 
 
 class TestInvalidMetadata(unittest.TestCase):
-    """Invalid Metadata test cases."""
+    """Invalid metadata test cases."""
 
     def test_empty_file(self):
-        """Check if a file is empty."""
+        """Check if an exception is raised if the file is empty."""
         settings = get_settings(
             PANDOC_EXTENSIONS=PANDOC_EXTENSIONS, PANDOC_ARGS=PANDOC_ARGS
         )
@@ -34,7 +36,7 @@ class TestInvalidMetadata(unittest.TestCase):
         self.assertEqual("Could not find metadata. File is empty.", message)
 
     def test_non_empty_file_no_metadata(self):
-        """Check if a file has no metadata."""
+        """Check if an exception is raised if the file has no metadata."""
         settings = get_settings(
             PANDOC_EXTENSIONS=PANDOC_EXTENSIONS, PANDOC_ARGS=PANDOC_ARGS
         )
@@ -50,7 +52,7 @@ class TestInvalidMetadata(unittest.TestCase):
         self.assertEqual("Could not find metadata header '---'.", message)
 
     def test_metadata_start_with_leading_spaces(self):
-        """Check if metadata block starting with leading spaces throws an exception."""
+        """Check that a metadata block start with leading spaces throws an exception."""
         settings = get_settings(
             PANDOC_EXTENSIONS=PANDOC_EXTENSIONS, PANDOC_ARGS=PANDOC_ARGS
         )
@@ -60,7 +62,7 @@ class TestInvalidMetadata(unittest.TestCase):
             TEST_CONTENT_PATH, "metadata_start_with_leading_spaces.md"
         )
 
-        # Metadata staring --- should not have leading spaces
+        # Metadata starting --- should not have leading spaces
         with self.assertRaises(Exception) as context_manager:
             pandoc_reader.read(source_path)
 
@@ -68,7 +70,7 @@ class TestInvalidMetadata(unittest.TestCase):
         self.assertEqual("Could not find metadata header '---'.", message)
 
     def test_metadata_block_end_with_leading_spaces(self):
-        """Check if metadata block ending with leading spaces throws an exception."""
+        """Check that a metadata block end with leading spaces throws an exception."""
         settings = get_settings(
             PANDOC_EXTENSIONS=PANDOC_EXTENSIONS, PANDOC_ARGS=PANDOC_ARGS
         )
@@ -86,7 +88,7 @@ class TestInvalidMetadata(unittest.TestCase):
         self.assertEqual("Could not find end of metadata block.", message)
 
     def test_no_metadata_block_end(self):
-        """Check if the metadata block ends."""
+        """Check that an exception is thrown if the metadata block does not end."""
         settings = get_settings(
             PANDOC_EXTENSIONS=PANDOC_EXTENSIONS, PANDOC_ARGS=PANDOC_ARGS
         )
@@ -102,13 +104,13 @@ class TestInvalidMetadata(unittest.TestCase):
         self.assertEqual("Could not find end of metadata block.", message)
 
     def test_invalid_metadata_block_end(self):
-        """Check if the metadata block end is wrong."""
+        """Check that an exception is thrown if the metadata block end is wrong."""
         settings = get_settings(
             PANDOC_EXTENSIONS=PANDOC_EXTENSIONS, PANDOC_ARGS=PANDOC_ARGS
         )
 
         pandoc_reader = PandocReader(settings)
-        source_path = os.path.join(TEST_CONTENT_PATH, "no_metadata_end.md")
+        source_path = os.path.join(TEST_CONTENT_PATH, "wrong_metadata_end.md")
 
         # Metadata blocks should end with '___' or '...' if not it should fail
         with self.assertRaises(Exception) as context_manager:
