+++ pyproject.toml
@@ -30,19 +30,20 @@ classifiers = [
 [tool.poetry.dependencies]
 python = ">=3.6.2,<4.0"
 pelican = ">=4.5"
-markdown = {version = ">=3.2", optional = true}
+markdown = {version = "<=3.3.4", optional = true}
 markdown-word-count = "^0.0.1"
 pyyaml = "^5.3.1"
 beautifulsoup4 = "^4.9.3"
+"ruamel.yaml" = "^0.17.21"
 
 [tool.poetry.dev-dependencies]
-black = {version = "^21.4b2", allow-prereleases = true}
-flake8 = "^3.9"
+black = {version = "*", allow-prereleases = true}
+flake8 = "^4.0.1"
 flake8-black = "^0.2.0"
 invoke = "^1.3"
 isort = "^5.4"
 livereload = "^2.6"
-markdown = "^3.2"
+markdown = "<=3.3.4"
 pytest = "^6.0"
 pytest-cov = "^2.8"
 pytest-pythonpath = "^0.7.3"
